renamed AuthContext to AuthProvider
