+++ components/auth/SignUp.tsx
@@ -16,7 +16,7 @@ import {
   SelectTrigger,
   SelectValue,
 } from "~/components/ui/select";
-import { useAuthentication } from "~/context/AuthContext";
+import { useAuthentication } from "~/providers/AuthProvider";
 import FormInput from "~/components/form/FormInput";
 import { Mail } from "lucide-react-native";
 import { signUpSchema } from "~/types/zod";
